+++ lib/sigepweb.dart
@@ -1,7 +1,8 @@
+/// Main package library
 library sigepweb;
 
+/// Export files
 export 'src/core.dart';
 export 'src/utils.dart';
 export 'src/constants.dart';
-export 'src/preco_prazo.dart';
 export 'src/models/contrato.dart';